Added LoadingBox to shared styledComponents
