+++ src/components/Shared/styledComponents/styledComponents.ts
@@ -1,6 +1,17 @@
-import { CircularProgress } from '@mui/material';
+import { Box, CircularProgress } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
 export const LoadingIndicator = styled(CircularProgress)(({ theme }) => ({
   margin: theme.spacing(0, 1, 0, 0),
+}));
+
+export const LoadingBox = styled(Box)(({ theme }) => ({
+  backgroundColor: theme.palette.cruGrayLight.main,
+  height: 300,
+  minWidth: 700,
+  margin: 'auto',
+  padding: 4,
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
 }));
